Make list of policies on UserDetails clickable. Fix Warnings. Add latest assets
